refactor(auth): move login logic into LoginForm component and simplify Login page
