+++ src/features/auth/components/login-form/index.ts
@@ -1,0 +1,1 @@
+export { default as LoginForm } from "./LoginForm";
